% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openblas.R
\name{get_num_threads}
\alias{get_num_threads}
\title{Get the number of allowed OpenBlas threads for current R session.}
\usage{
get_num_threads()
}
\description{
Get the number of allowed OpenBlas threads for current R session.
}
\examples{
get_num_threads()
set_num_threads(2)
get_num_threads()

}
