% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openblas.R
\name{set_num_threads}
\alias{set_num_threads}
\title{Wrapper to set the number of OpenBlas threads from within an R session, via 
openblas_set_num_threads(), for implicit parallelization. Alters the number 
of threads employed by OpenBlas for that session and child sessions.}
\usage{
set_num_threads(threads = parallel::detectCores())
}
\arguments{
\item{threads}{Maximum number of threads to use for OpenBlas operations}
}
\description{
Wrapper to set the number of OpenBlas threads from within an R session, via 
openblas_set_num_threads(), for implicit parallelization. Alters the number 
of threads employed by OpenBlas for that session and child sessions.
}
\examples{
mat = matrix(seq_len(5e7),5e4,1e3)

set_num_threads(1)
system.time( crossprod(mat,mat) )

openblas.set.num.threads(2)
system.time( crossprod(mat,mat) )

openblas.set.num.threads(4)
system.time( crossprod(mat,mat) )

}
